move str handling to utils
